+++ Pluminus/NC1App.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct NC1App: App {
     var body: some Scene {
         WindowGroup {
-            MainContentView()
+            MainView()
         }
     }
 }
